requirements.txt
strands-agents>=1.35.0
strands-agents-tools>=0.2.0
